+++ pkg/rtprtcp/rtp.go
@@ -123,6 +123,9 @@ func ParseRTPPacket(b []byte) (pkt RTPPacket, err error) {
 }
 
 // 比较序号的值，内部处理序号翻转问题，见单元测试中的例子
+// @return  0 a和b相等
+//          1 a大于b
+//         -1 a小于b
 func CompareSeq(a, b uint16) int {
 	if a == b {
 		return 0